pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "Flask>=3.1.1",
    "Flask-SQLAlchemy>=3.1.1",
    "Flask-Login>=0.6.3",
    "Flask-Migrate>=4.0.7",
    "python-dotenv>=1.0.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "google-generativeai>=0.8.5",
    "numpy>=2.2.6",
    "scikit-learn>=1.6.1",
    "defusedcsv>=2.0.0",
    "PyPDF2>=3.0.1",
    "python-docx>=1.1.2",
    "requests>=2.31.0",
    "Flask-Talisman>=1.1.0",
]
